refactor: moved hashStatements to Crypto
