+++ src/crypto/Crypto.ts
@@ -18,7 +18,7 @@ import {
   u8aToString,
   u8aToU8a,
 } from '@polkadot/util'
-import { signatureVerify } from '@polkadot/util-crypto'
+import { blake2AsHex, signatureVerify } from '@polkadot/util-crypto'
 import blake2AsU8a from '@polkadot/util-crypto/blake2/asU8a'
 import naclDecrypt from '@polkadot/util-crypto/nacl/decrypt'
 import naclEncrypt from '@polkadot/util-crypto/nacl/encrypt'
@@ -212,4 +212,63 @@ export function decryptAsymmetricAsStr(
     coToUInt8(secretKeyA)
   )
   return result ? u8aToString(result) : false
+}
+
+export interface Hasher {
+  (value: string, nonce?: string): string
+}
+
+export interface HashingOptions {
+  nonces?: Record<string, string>
+  nonceGenerator?: (key: string) => string
+  hasher?: Hasher
+}
+
+const saltedBlake2b256: Hasher = (value, nonce) =>
+  blake2AsHex((nonce || '') + value, 256)
+
+/**
+ * Configurable computation of salted or unsalted hashes over an array or map of statements. Can be used to validate/reproduce salted hashes
+ * by means of an optional nonce map.
+ *
+ * @param statements A map or array of statement strings to be hashed. Keys will be used to look up nonces or as input for the `nonceGenerator`.
+ * @param options Optional hasher arguments.
+ * @param options.nonces An optional map of nonces. If present, it must comprise all keys of `statements`, as those will be used map nonces to statements.
+ * @param options.nonceGenerator An optional nonce generator. Will be used if `options.nonces` is not defined to generate a (new) nonce for each statement. The statement key is passed as its first argument. If no `nonces` or `nonceGenerator` are given this function returns unsalted hashes.
+ * @param options.hasher The hasher to be used. Computes a hash from a statement and an optional nonce. Required but defaults to 256 bit blake2 over `${nonce}${statement}`.
+ * @returns An array of objects for each statement which contain the hash, nonce and statement key.
+ */
+export function hashStatements(
+  statements: string[] | Record<string, string>,
+  options: HashingOptions = {}
+): Array<{ key: string; hash: string; nonce?: string }> {
+  // apply default hasher if undefined
+  const { hasher, nonces, nonceGenerator } = {
+    ...options,
+    hasher: options.hasher || saltedBlake2b256,
+  }
+  // getNonce <- if nonces map then take from map, else use nonceGenerator function
+  // if no nonceGenerator given, unsalted hashes are returned
+  const getNonce: HashingOptions['nonceGenerator'] =
+    typeof nonces === 'object'
+      ? (key) => {
+          const nonce = nonces[key]
+          // nonce map must be exhaustive
+          // TODO: use sdk error module
+          if (!nonce)
+            throw new Error(
+              `no nonce given for statement with nonce key ${key}`
+            )
+          return nonce
+        }
+      : nonceGenerator
+  // map over statements, use key to retrieve nonce, then hash statement & nonce
+  return Object.entries(statements).map(([key, statement]) => {
+    const nonce = getNonce ? getNonce(key) : undefined
+    return {
+      key,
+      nonce,
+      hash: hasher(statement, nonce),
+    }
+  })
 }
